Refactor: splitted engine.h to finalEngine.h and viewportEngine.h
