+++ extern/usdhydra/engine.h
@@ -50,37 +50,6 @@ protected:
   pxr::UsdStageRefPtr stage;
 };
 
-class FinalEngine : public Engine {
-public:
-  using Engine::Engine;
-  void sync(BL::Depsgraph &b_depsgraph, BL::Context &b_context, pxr::HdRenderSettingsMap &renderSettings) override;
-  void render(BL::Depsgraph &b_depsgraph);
-
-private:
-  void renderGL(BL::Depsgraph &b_depsgraph);
-  void renderLite(BL::Depsgraph &b_depsgraph);
-  void getResolution(BL::RenderSettings b_render, int &width, int &height);
-  void updateRenderResult(std::map<std::string, std::vector<float>> &render_images, const std::string &layerName, int width, int height);
-  void notifyStatus(float progress, const std::string &title, const std::string &info);
-};
-
-class ViewportEngine : public Engine {
-public:
-  using Engine::Engine;
-  void sync(BL::Depsgraph &b_depsgraph, BL::Context &b_context, pxr::HdRenderSettingsMap &renderSettings) override;
-  void viewDraw(BL::Depsgraph &b_depsgraph, BL::Context &b_context);
-
-private:
-  void notifyStatus(const std::string &title, const std::string &info, bool redraw);
-
-private:
-  std::unique_ptr<pxr::UsdImagingGLEngine> imagingGLEngine;
-  pxr::UsdImagingGLRenderParams renderParams;
-  std::chrono::time_point<std::chrono::steady_clock> timeBegin;
-};
-
-PyObject *addPythonSubmodule_engine(PyObject *mod);
-
 template <typename T>
 float Engine::getRendererPercentDone(T &renderer)
 {
@@ -99,5 +68,7 @@ inline pxr::UsdStageRefPtr Engine::getStage()
 {
   return stage;
 }
+
+PyObject *addPythonSubmodule_engine(PyObject *mod);
 
 }   // namespace usdhydra
