+++ extern/usdhydra/finalEngine.cpp
@@ -10,7 +10,7 @@
 
 #include "glog/logging.h"
 
-#include "engine.h"
+#include "finalEngine.h"
 #include "utils.h"
 #include "sceneDelegate/scene.h"
