+++ extern/usdhydra/viewportEngine.cpp
@@ -11,7 +11,7 @@
 
 #include "glog/logging.h"
 
-#include "engine.h"
+#include "viewportEngine.h"
 #include "utils.h"
 
 using namespace std;
